Add AppLayout to NotFoundView
